# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Klar API",
    "description": "German bureaucratic mail \u2192 structured obligations + deadlines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/letters/upload": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Upload a Brief (image or PDF)",
        "description": "Multipart upload of a German letter. Returns immediately with the new `letter_id`; the heavy AI extraction runs LATER via `GET /api/letters/{letter_id}/process` (SSE).\n\nAccepts: JPEG / PNG / WebP / HEIC / PDF, up to 10 MB. Server verifies magic bytes against the declared `Content-Type` \u2014 defends against binaries renamed `.jpg`.",
        "operationId": "upload_letter_api_letters_upload_post",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 8
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_letter_api_letters_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "`LETTER_EMPTY_UPLOAD`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`LETTER_TOO_LARGE` (>10 MB).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`LETTER_UNSUPPORTED_TYPE` (bad MIME), `LETTER_CORRUPT_FILE` (no magic match), or `LETTER_MIME_MISMATCH` (declared \u2260 detected).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters/{letter_id}/extract": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Synchronous extraction (no streaming)",
        "description": "Runs ONE structured Qwen call and returns the full LetterResponse. Long-form fields (`explanation`, `response_draft`, `checklist`, `citations`) are NOT populated here \u2014 use the SSE `/process` route for those. Used by clients that don't support EventSource.",
        "operationId": "extract_letter_api_letters__letter_id__extract_post",
        "parameters": [
          {
            "name": "letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`LETTER_NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`LETTER_FILE_MISSING`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "`EXTRACTION_FAILED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "List the current user's letters",
        "description": "Sorted by `created_at` DESC. Optional filters:\n- `?status=uploaded|processing|completed|error`\n- `?category=health_insurance|tax|immigration|...` (see DocumentCategory enum)\n\nEmpty-string values for `status` / `category` are treated as 'no filter' \u2014 safe to bind directly to React `useState('')`.",
        "operationId": "list_letters_api_letters_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LetterListItem"
                  },
                  "title": "Response List Letters Api Letters Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unknown status or category value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters/{letter_id}/process": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "SSE stream: full extraction + long-form generation",
        "description": "Server-Sent Events stream that orchestrates the AI pipeline. The response is `text/event-stream`; each frame has `event: <type>` and `data: <json>` lines.\n\nEvent sequence (each event JSON shape is documented in schemas.py):\n1. `ocr_result` \u2014 verbatim German text (one frame)\n2. `classification` \u2014 document type + category + agency\n3. `risk_score` \u2014 integer 0\u2013100 + label\n4. `deadline` \u2014 earliest action's deadline + days remaining\n5. `consequence` \u2014 what happens if the user does nothing\n6. `explanation` \u2014 streaming, multiple frames, one chunk per frame\n7. `response_draft` \u2014 streaming, conditional (only if any action needs a reply); ALWAYS in German regardless of `?lang`\n8. `checklist` \u2014 array of items the user must gather\n9. `citations` \u2014 RAG corpus hits used to ground the extraction\n10. `done` \u2014 final frame, `{ letter_id }`. Frontend should close the EventSource.\n\nOn any failure, a single `error` frame is sent with the standard Klar error envelope (`code` + `message`), then the stream closes.\n\n**Frontend:** open with `new EventSource(url, { withCredentials: true })` so the `klar_session` cookie travels.",
        "operationId": "process_letter_api_letters__letter_id__process_get",
        "parameters": [
          {
            "name": "letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Letter Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 8
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream \u2014 see description for event sequence.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`LETTER_NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/letters/{letter_id}": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "Get a single letter with all its actions",
        "description": "Returns the full `LetterResponse` including denormalized `risk_score`, `deadline_date`, every `ActionItem` (with `evidence_span`, `deadline_confidence`, `steps`), plus any long-form fields (`explanation`, `response_draft`, `checklist`, `citations`) the `/process` pipeline has already filled in.",
        "operationId": "get_letter_api_letters__letter_id__get",
        "parameters": [
          {
            "name": "letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`LETTER_NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_letter_api_letters_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_letter_api_letters_upload_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable identifier \u2014 see docs/06-api-contract.md."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Localized, user-facing copy."
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Every non-2xx response from this API has this shape.\n\nFrontend code should switch on `code` (stable identifier), display\n`message` to the user, and inspect `details` for field-level\ninformation on `VALIDATION_ERROR`s."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LetterListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "letter_type": {
            "type": "string",
            "title": "Letter Type"
          },
          "category": {
            "$ref": "#/components/schemas/DocumentCategory"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "deadline_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Date"
          },
          "status": {
            "$ref": "#/components/schemas/LetterStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "letter_type",
          "category",
          "risk_score",
          "status",
          "created_at"
        ],
        "title": "LetterListItem",
        "description": "Compact shape used by GET /api/letters list endpoint."
      },
      "LetterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "institution": {
            "type": "string",
            "title": "Institution"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "letter_type": {
            "type": "string",
            "title": "Letter Type"
          },
          "category": {
            "$ref": "#/components/schemas/DocumentCategory"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "deadline_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Date"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "response_draft": {
            "type": "string",
            "title": "Response Draft",
            "default": ""
          },
          "checklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checklist"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "consequence": {
            "type": "string",
            "title": "Consequence",
            "default": ""
          },
          "status": {
            "$ref": "#/components/schemas/LetterStatus"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "actions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Actions"
          },
          "extraction_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extraction Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "institution",
          "document_type",
          "letter_type",
          "category",
          "summary",
          "language",
          "risk_score",
          "status",
          "created_at"
        ],
        "title": "LetterResponse"
      },
...
      "LetterUploadResponse": {
        "properties": {
          "letter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Letter Id"
          }
        },
        "type": "object",
        "required": [
          "letter_id"
        ],
        "title": "LetterUploadResponse"
      },
...
    }
  }
}